# backend/api/openapi.yml
openapi: 3.0.3
info:
  title: Warehouse Management System API
  description: API для системы управления складом.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке
          example: "invalid input parameter"
        details:
          type: string
          description: Детали ошибки (обычно при валидации JSON)
          nullable: true
    Paging:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
...
    CreateUserRequest:
      type: object
      required: [username, email, full_name, role]
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
          enum: [admin, worker]
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
        is_active:
          type: boolean
    PagedUsers:
      type: object
      properties:
        paging:
          $ref: '#/components/schemas/Paging'
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
...
  responses:
    BadRequest:
      description: Неверные параметры запроса (ErrInvalidInput или ошибка валидации JSON)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "invalid input parameter"
    Unauthorized:
      description: Не авторизован (ErrUnauthorized)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "unauthorized"
    Forbidden:
      description: Нет прав доступа (недостаточно прав для роли)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Access denied: insufficient permissions"
    NotFound:
      description: Ресурс не найден (ErrNotFound)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "not found"
    Conflict:
      description: Конфликт данных (ErrAlreadyExists или ErrInsufficientStock)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "already exists"
    InternalError:
      description: Внутренняя ошибка сервера (ErrInternal или panic)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "internal server error"
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: Получить список пользователей
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedUsers'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Users]
      summary: Создать пользователя
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        201:
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        409:
          $ref: '#/components/responses/Conflict'
        500:
          $ref: '#/components/responses/InternalError'
  /api/users/{id}:
    delete:
      tags: [Users]
      summary: Удалить пользователя
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Пользователь удален
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'
...